Small ESP & GOBJECTS update
Made items distance at the end of the name and added centering on items / task esp like done with players.

Updated game functions to remove stuff not needed and fix gun data to not do defaults.

Removed config files so we dont error when adding new settings and them not existing, clicking save on settings will generate the file.